+++ src/main.ts
@@ -10,6 +10,28 @@ import { OpenArticleModal } from "./commands/openArticles";
 import { wikipediaIcon } from "./utils/wikipediaIcon";
 import { CreateArticleNoteModal } from "./commands/createArticleNotes";
 
+// Missing Commons and Wikidata
+export type Wiki =
+	| "Wikipedia"
+	| "Wiktionary"
+	| "Wikibooks"
+	| "Wikiquote"
+	| "Wikiversity"
+	| "Wikivoyage"
+	| "Wikisource"
+	| "Wikinews";
+
+export const wikilist: Wiki[] = [
+	"Wikipedia",
+	"Wiktionary",
+	"Wikibooks",
+	"Wikiquote",
+	"Wikiversity",
+	"Wikivoyage",
+	"Wikisource",
+	"Wikinews",
+];
+
 export default class WikipediaSearchPlugin extends Plugin {
 	settings: WikipediaSearchSettings;
 
@@ -19,33 +41,49 @@ export default class WikipediaSearchPlugin extends Plugin {
 		await this.loadSettings();
 
 		this.addCommand({
-			id: "link-article",
-			name: "Link Article",
-			editorCallback: (editor: Editor) => {
-				new LinkArticleModal(this.app, this.settings, editor).open();
-			},
+			id: "link-wikipedia-article",
+			name: "Link Wikipedia Article",
+			editorCallback: (editor: Editor) =>
+				new LinkArticleModal(this.app, this.settings, "Wikipedia", editor).open(),
 		});
 
 		this.addCommand({
-			id: "open-article",
-			name: "Open Article",
-			callback: () => new OpenArticleModal(this, this.settings).open(),
+			id: "open-wikipedia-article",
+			name: "Open Wikipedia Article",
+			callback: () => new OpenArticleModal(this.app, this.settings, "Wikipedia", this).open(),
 		});
 
 		this.addCommand({
-			id: "create-article-note",
-			name: "Create Article Note",
+			id: "create-wikipedia-article-note",
+			name: "Create Wikipedia Article Note",
 			callback: () => {
 				if (this.settings.templates.filter((template) => template.createNote).length === 0) {
 					new Notice("To use this command you have to create a note template first!");
 					return;
 				}
-				new CreateArticleNoteModal(this.app, this.settings).open();
+				new CreateArticleNoteModal(this.app, this.settings, "Wikipedia").open();
 			},
 		});
 
 		addIcon("wikipedia", wikipediaIcon);
-		this.addRibbonIcon("wikipedia", "Open Article", () => new OpenArticleModal(this, this.settings).open());
+		this.addRibbonIcon("wikipedia", "Open Article", () =>
+			new OpenArticleModal(this.app, this.settings, "Wikipedia", this).open()
+		);
+
+		for (const wiki of wikilist) {
+			this.addCommand({
+				id: `link-${wiki.toLowerCase()}-article`,
+				name: `Link ${wiki} Article`,
+				editorCallback: (editor: Editor) =>
+					new LinkArticleModal(this.app, this.settings, wiki, editor).open(),
+			});
+
+			this.addCommand({
+				id: `open-${wiki.toLowerCase()}-article`,
+				name: `Open ${wiki} Article`,
+				callback: () => new OpenArticleModal(this.app, this.settings, wiki, this).open(),
+			});
+		}
 
 		this.addSettingTab(new WikipediaSearchSettingTab(this.app, this));
 	}